+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to Pipelines will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- `--verbatim` option to only output from pipeline, not pipelines
+
 ## [0.0.2] - 2018-01-11
 ### Added
 - Brace glob pattern in pipelines